Remove Source Generation
Despite the convenience of source generation, it's evident that its current state does not align with an efficient and satisfactory development experience. This discrepancy stems from various factors including inadequate support from major IDEs/editors, confinement to netstandard2.0, and unexpected behavior exhibited by commands such as `dotnet build`. As a result, considering its present condition, the suitability of sgen for integration into this project is questionable and is left undesired.