Removed design flaw which could cause lost tasks
Take the following scenario inside of the ConcurrentStack
```
Task 1
Task 2
Task 3
```
If task 2 finishes before task 3, task 3 is removed. This means there could still be active connections when `StopAsync` is called, however those connections will still remain open when the server has closed. Worse, they could throw when `TcpListener.Stop` is called.
The solution here is to switch to a dictionary keyed by Ulid's. Since nearly all Ulid's are unique by design, the chance of a collision is insanely low. Because of the extremely low odds, I've opted to skip using the ConcurrentDictionary. If this choice proves to be foolish, simply modify `ListenForConnectionsAsync` to use TryAdd and continue trying until succeeding.